// src/app/api/models/candidate-registration-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CandidateRegistrationDto {
  alternateContactNumber?: string;
  contactNumber: string;
  disability?: string;
  firstName: string;
  gender?: string;
  lastName: string;
  race?: string;
  username: string;
}